Fix sku limit in product variant list
