+++ src/products/components/ProductVariants/ProductVariants.tsx
@@ -5,11 +5,13 @@ import Hidden from "@material-ui/core/Hidden";
 import { fade } from "@material-ui/core/styles/colorManipulator";
 import TableCell from "@material-ui/core/TableCell";
 import Typography from "@material-ui/core/Typography";
+import Alert from "@saleor/components/Alert/Alert";
 import CardTitle from "@saleor/components/CardTitle";
 import Checkbox from "@saleor/components/Checkbox";
 import LinkChoice from "@saleor/components/LinkChoice";
 import Money from "@saleor/components/Money";
 import ResponsiveTable from "@saleor/components/ResponsiveTable";
+import { RefreshLimits_shop_limits } from "@saleor/components/Shop/types/RefreshLimits";
 import { SingleAutocompleteChoiceType } from "@saleor/components/SingleAutocompleteSelectField";
 import Skeleton from "@saleor/components/Skeleton";
 import {
@@ -18,6 +20,7 @@ import {
 } from "@saleor/components/SortableTable";
 import TableHead from "@saleor/components/TableHead";
 import { makeStyles } from "@saleor/theme";
+import { isLimitReached } from "@saleor/utils/limits";
 import React from "react";
 import { FormattedMessage, IntlShape, useIntl } from "react-intl";
 
@@ -81,6 +84,9 @@ const useStyles = makeStyles(
       colSku: {
         width: 200
       }
+    },
+    alert: {
+      margin: theme.spacing(3, 3, 0)
     },
     colGrab: {
       width: 60
@@ -179,8 +185,8 @@ function getAvailabilityLabel(
 }
 
 interface ProductVariantsProps extends ListActions, ChannelProps {
-  addButtonDisabled: boolean;
   disabled: boolean;
+  limits: RefreshLimits_shop_limits;
   product: ProductDetails_product;
   variants: ProductDetails_product_variants[];
   onVariantReorder: ReorderAction;
@@ -194,8 +200,8 @@ const numberOfColumns = 7;
 
 export const ProductVariants: React.FC<ProductVariantsProps> = props => {
   const {
-    addButtonDisabled,
     disabled,
+    limits,
     variants,
     product,
     onRowClick,
@@ -215,6 +221,7 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
   const intl = useIntl();
   const [warehouse, setWarehouse] = React.useState<string>(null);
   const hasVariants = maybe(() => variants.length > 0, true);
+  const limitReached = isLimitReached(limits, "productVariants");
 
   return (
     <Card>
@@ -226,7 +233,7 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
         toolbar={
           hasVariants ? (
             <Button
-              disabled={addButtonDisabled}
+              disabled={limitReached}
               onClick={onVariantAdd}
               variant="text"
               color="primary"
@@ -239,7 +246,7 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
             </Button>
           ) : (
             <Button
-              disabled={addButtonDisabled}
+              disabled={limitReached}
               onClick={onVariantsAdd}
               variant="text"
               color="primary"
@@ -253,6 +260,17 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
           )
         }
       />
+
+      <Alert
+        className={classes.alert}
+        show={limitReached}
+        title={intl.formatMessage({
+          defaultMessage: "SKU limit reached",
+          description: "alert"
+        })}
+      >
+        <FormattedMessage defaultMessage="You have reached your SKU limit, you will be no longer able to add SKUs to your store. If you would like to up your limit, contact your administration staff about raising your limits." />
+      </Alert>
 
       {variants.length > 0 ? (
         <CardContent className={classes.warehouseSelectContainer}>
